Add a books section
I've added a books section. Two of the books are actually available to read. One of them just links to a GoodReads page for the book. I wasn't sure if it would be better to keep the books consistent and just add Amazon links or do it this way. I decided to go with the route that allowed people to actually read the books!